[ETCM-105] Pr comments
Refactor FastSyncIt tests
Properly close actor system in StateSyncSpec
