+++ src/test/scala/io/iohk/ethereum/blockchain/sync/SyncSchedulerSpec.scala
@@ -25,15 +25,16 @@ class SyncSchedulerSpec extends AnyFlatSpec with Matchers with EitherValues with
     val (scheduler, schedulerBlockchain, schedulerDb) = buildScheduler()
     val initialState = scheduler.initState(worldHash).get
     val (missingNodes, newState) = scheduler.getMissingNodes(initialState, 1)
-    assert(missingNodes.size == 1)
     val responses = prov.getNodes(missingNodes)
-    assert(responses.size == 1)
     val result = scheduler.processResponses(newState, responses)
-    assert(result.isRight)
     val (newRequests, state) = scheduler.getMissingNodes(result.right.value._1, 1)
+    scheduler.persistBatch(state, 1)
+
+    assert(missingNodes.size == 1)
+    assert(responses.size == 1)
+    assert(result.isRight)
     assert(newRequests.isEmpty)
     assert(state.numberOfPendingRequests == 0)
-    scheduler.persistBatch(state, 1)
     assert(schedulerDb.storages.nodeStorage.get(missingNodes.head).isDefined)
   }
 
@@ -45,15 +46,14 @@ class SyncSchedulerSpec extends AnyFlatSpec with Matchers with EitherValues with
     val (scheduler, schedulerBlockchain, schedulerDb) = buildScheduler()
     val initState = scheduler.initState(worldHash).get
     val state1 = exchangeSingleNode(initState, scheduler, prov).right.value
-    assert(state1.numberOfPendingRequests > 0)
-
     val state2 = exchangeSingleNode(state1, scheduler, prov).right.value
-    assert(state2.numberOfPendingRequests > 0)
-
     val state3 = exchangeSingleNode(state2, scheduler, prov).right.value
+    scheduler.persistBatch(state3, 1)
+
+    assert(state1.numberOfPendingRequests > 0)
+    assert(state2.numberOfPendingRequests > 0)
     // only after processing third result request is finalized as code and storage of account has been retrieved
     assert(state3.numberOfPendingRequests == 0)
-    scheduler.persistBatch(state3, 1)
     // 1 leaf node + 1 code + 1 storage
     assert(schedulerDb.dataSource.storage.size == 3)
   }
@@ -69,34 +69,34 @@ class SyncSchedulerSpec extends AnyFlatSpec with Matchers with EitherValues with
     )
     val (scheduler, schedulerBlockchain, schedulerDb) = buildScheduler()
     val initState = scheduler.initState(worldHash).get
-    val state1 = exchangeSingleNode(initState, scheduler, prov).right.value
-    // 1 non finalized request for branch node + 2 non finalized request for leaf nodes
-    assert(state1.numberOfPendingRequests == 3)
     assert(schedulerDb.dataSource.storage.isEmpty)
-
+    val state1 = exchangeSingleNode(initState, scheduler, prov).right.value
     val state2 = exchangeSingleNode(state1, scheduler, prov).right.value
-    // 1 non finalized request for branch node + 2 non finalized requests for leaf nodes + 2 non finalized requests for code and
-    // storage
-    assert(state2.numberOfPendingRequests == 5)
-
     val state3 = exchangeSingleNode(state2, scheduler, prov).right.value
     val state4 = exchangeSingleNode(state3, scheduler, prov).right.value
     val state5 = scheduler.persistBatch(state4, 1)
     // finalized leaf node i.e state node + storage node + code
     assert(schedulerDb.dataSource.storage.size == 3)
+    val state6 = exchangeSingleNode(state5, scheduler, prov).right.value
+    val state7 = exchangeSingleNode(state6, scheduler, prov).right.value
+    val state8 = exchangeSingleNode(state7, scheduler, prov).right.value
+    val state9 = scheduler.persistBatch(state8, 1)
+
+    // 1 non finalized request for branch node + 2 non finalized request for leaf nodes
+    assert(state1.numberOfPendingRequests == 3)
+
+    // 1 non finalized request for branch node + 2 non finalized requests for leaf nodes + 2 non finalized requests for code and
+    // storage
+    assert(state2.numberOfPendingRequests == 5)
 
     // 1 non finalized request for branch node + 1 non finalized request for leaf node
     assert(state5.numberOfPendingRequests == 2)
 
-    val state6 = exchangeSingleNode(state5, scheduler, prov).right.value
     // 1 non finalized request for branch node + 1 non finalized request for leaf node + 2 non finalized request for code and storage
     assert(state6.numberOfPendingRequests == 4)
 
-    val state7 = exchangeSingleNode(state6, scheduler, prov).right.value
-    val state8 = exchangeSingleNode(state7, scheduler, prov).right.value
     // received code and storage finalized remaining leaf node, and branch node
     assert(state8.numberOfPendingRequests == 0)
-    val state9 = scheduler.persistBatch(state8, 1)
     // 1 branch node + 2 leaf nodes + 4 code and storage data
     assert(state9.numberOfPendingRequests == 0)
     assert(schedulerDb.dataSource.storage.size == 7)
@@ -114,19 +114,22 @@ class SyncSchedulerSpec extends AnyFlatSpec with Matchers with EitherValues with
     val (scheduler, schedulerBlockchain, schedulerDb) = buildScheduler()
     val initState = scheduler.initState(worldHash).get
     val state1 = exchangeSingleNode(initState, scheduler, prov).right.value
-    // 1 non finalized request for branch node + 2 non finalized request for leaf nodes
-    assert(state1.numberOfPendingRequests == 3)
     val (allMissingNodes1, state2) = scheduler.getAllMissingNodes(state1)
-    assert(allMissingNodes1.size == 2)
     val allMissingNodes1Response = prov.getNodes(allMissingNodes1)
     val state3 = scheduler.processResponses(state2, allMissingNodes1Response).right.value._1
     val (allMissingNodes2, state4) = scheduler.getAllMissingNodes(state3)
-    assert(allMissingNodes2.size == 2)
     val allMissingNodes2Response = prov.getNodes(allMissingNodes2)
     val state5 = scheduler.processResponses(state4, allMissingNodes2Response).right.value._1
     val remaingNodes = state5.numberOfPendingRequests
-    assert(remaingNodes == 0)
     val state6 = scheduler.persistBatch(state5, 1)
+
+    // 1 non finalized request for branch node + 2 non finalized request for leaf nodes
+    assert(state1.numberOfPendingRequests == 3)
+    assert(allMissingNodes1.size == 2)
+
+    assert(allMissingNodes2.size == 2)
+
+    assert(remaingNodes == 0)
     // 1 branch node + 2 leaf node + 1 code + 1 storage (code and storage are shared by 2 leaf nodes)
     assert(schedulerDb.dataSource.storage.size == 5)
   }
@@ -162,9 +165,10 @@ class SyncSchedulerSpec extends AnyFlatSpec with Matchers with EitherValues with
     val (firstMissing, state1) = scheduler.getMissingNodes(initState, 1)
     val firstMissingResponse = prov.getNodes(firstMissing)
     val result1 = scheduler.processResponse(state1, firstMissingResponse.head)
-    assert(result1.isRight)
     val stateAfterReceived = result1.right.value
     val result2 = scheduler.processResponse(stateAfterReceived, firstMissingResponse.head)
+
+    assert(result1.isRight)
     assert(result2.isLeft)
     assert(result2.left.value == AlreadyProcessedItem)
   }
